Add conditionals in case systems are already cleaned up.

+++ cleanup.sh
@@ -7,13 +7,14 @@ function_name=$1
 function cleanup_master() {
     # Destroy openstack and delete images
     if [ -d "/opt/kolla" ]; then
-        cd /opt/kolla || exi
+        cd /opt/kolla || exit
         python3 -m venv venv
         source venv/bin/activate
         kolla-ansible -i multinode destroy --yes-i-really-really-mean-it --include-images
         deactivate
         sudo pip3 uninstall -qy python-openstackclient
     fi
+    
     # Cleanup refstack
     cd ~ || exit
     sudo rm -fr refstack-client
@@ -70,10 +71,12 @@ function cleanup_master() {
 function cleanup_nodes() {
     # This will remove: all stopped containers, all networks not used by at least one container,
     # all volumes not used by at least one container, all dangling images, all build cache
-    sudo docker rm -f "$(sudo docker ps -aq)"
-    sudo docker system prune --volumes --force -a
+    if [[  $(sudo docker ps) ]]; then
+        sudo docker rm -f "$(sudo docker ps -aq)"
+        sudo docker system prune --volumes --force -a
+    fi
 
-    # Cleanup up ssh know hosts
+    # Cleanup up ssh know_hosts
     rm -f ~/.ssh/known_hosts
 
     # Cleanup kolla dirs
@@ -85,46 +88,48 @@ function cleanup_nodes() {
 
 function cleanup_storage_nodes() {
     # ceph services cleanup
-    ceph_services=$(sudo systemctl |grep ceph |grep "\.service" |awk '{print $1}')
-    for ceph_service in $ceph_services
-    do
-        sudo systemctl stop "$ceph_service"
-        sleep 1
-        sudo systemctl disable "$ceph_service"
-    done
+    if [[  $(sudo ceph -s) ]]; then
+        ceph_services=$(sudo systemctl |grep ceph |grep "\.service" |awk '{print $1}')
+        for ceph_service in $ceph_services
+        do
+            sudo systemctl stop "$ceph_service"
+            sleep 1
+            sudo systemctl disable "$ceph_service"
+        done
 
-    sudo rm -fr /etc/systemd/system/ceph*
-    sudo rm -fr /etc/systemd/system/*/ceph*
-    sudo rm -fr /lib/systemd/system/ceph*
+        sudo rm -fr /etc/systemd/system/ceph*
+        sudo rm -fr /etc/systemd/system/*/ceph*
+        sudo rm -fr /lib/systemd/system/ceph*
 
-    sudo systemctl daemon-reload
-    sudo systemctl reset-failed
+        sudo systemctl daemon-reload
+        sudo systemctl reset-failed
 
-    array=()
-    while IFS='' read -r line; do array+=("$line"); done < <(lsblk -dp | grep -o '^/dev[^ ]*')
-    new_array=()
-    working_dir=$(df | grep '^/' | awk '{print $1}')
-    for value in "${array[@]}"
-    do
-        [[ $working_dir =~ $value ]] || new_array+=("$value")
-    done
-    array=("${new_array[@]}")
-    unset new_array
+        array=()
+        while IFS='' read -r line; do array+=("$line"); done < <(lsblk -dp | grep -o '^/dev[^ ]*')
+        new_array=()
+        working_dir=$(df | grep '^/' | awk '{print $1}')
+        for value in "${array[@]}"
+        do
+            [[ $working_dir =~ $value ]] || new_array+=("$value")
+        done
+        array=("${new_array[@]}")
+        unset new_array
 
-    for device in "${array[@]}"
-    do
-        echo "INFO: Working working on device [$device]"
-        sudo umount "$device"
-        sudo sgdisk --zap-all -- "$device"
-        sudo partprobe || true
-        sudo wipefs -af "$device"
-    done
+        for device in "${array[@]}"
+        do
+            echo "INFO: Working working on device [$device]"
+            sudo umount "$device"
+            sudo sgdisk --zap-all -- "$device"
+            sudo partprobe || true
+            sudo wipefs -af "$device"
+        done
 
-    # remove all logical devices that use the /dev/mapper driver
-    sudo dmsetup remove_all
+        # remove all logical devices that use the /dev/mapper driver
+        sudo dmsetup remove_all
 
-    # Remove the ceph package
-    sudo rm -fr /etc/apt/sources.list.d/ceph.list
+        # Remove the ceph package
+        sudo rm -fr /etc/apt/sources.list.d/ceph.list
+    fi
 }
 
 $function_name
